+++ app/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     implementation(projects.feature.playlist)
     implementation(projects.feature.relaylist)
     implementation(projects.feature.mypage)
+    implementation(projects.feature.devmode)
     implementation(projects.feature.photocard)
     implementation(projects.feature.song)
