--- core/navigator/src/main/java/com/wepli/navigator/feature/mypage/NetworkLogRoute.kt
@@ -1,6 +1,0 @@
-package com.wepli.navigator.feature.mypage
-
-enum class NetworkLogRoute(val route: String) {
-    Main("network_main"),
-    Detail("network_detail"),
-}
