+++ data/src/main/kotlin/com/wepli/data/db/devmode/entity/ApiLogEntity.kt
@@ -3,9 +3,8 @@ package com.wepli.data.db.devmode.entity
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import debug.model.ApiLog
-import debug.model.ApiMethod
-import java.util.UUID
+import com.wepli.domain.devmode.model.ApiLog
+import com.wepli.domain.devmode.model.ApiMethod
 
 @Entity(
     tableName = "ApiLogs",
