+++ data/src/main/kotlin/com/wepli/data/di/repository/RepositoryModule.kt
@@ -15,7 +15,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import debug.repository.DebugApiLogRepository
+import com.wepli.domain.devmode.repository.DebugApiLogRepository
 import com.wepli.data.network.apilog.DebugApiLogRepositoryImpl
 import repository.youtube.YoutubeRepository
 import com.wepli.data.youtube.repository.YoutubeRepositoryImpl
