+++ data/src/main/kotlin/com/wepli/data/network/apilog/datasource/DebugApiLogLocalDatasource.kt
@@ -1,7 +1,6 @@
 package com.wepli.data.network.apilog.datasource
 
 import com.wepli.data.db.devmode.entity.ApiLogEntity
-import kotlinx.coroutines.flow.Flow
 
 interface DebugApiLogLocalDatasource {
