+++ domain/src/main/kotlin/com/wepli/domain/devmode/repository/DebugApiLogRepository.kt
@@ -1,0 +1,12 @@
+package com.wepli.domain.devmode.repository
+
+import com.wepli.domain.devmode.model.ApiLog
+import kotlinx.coroutines.flow.Flow
+
+interface DebugApiLogRepository {
+    suspend fun getLogs(count: Int): List<ApiLog>
+
+    suspend fun insertLog(log: ApiLog)
+
+    suspend fun findLogById(logId: Int): ApiLog?
+}
