+++ feature/devmode/src/main/kotlin/com/wepli/feature/devmode/navigation/NetworkLogRoute.kt
@@ -1,0 +1,6 @@
+package com.wepli.feature.devmode.navigation
+
+enum class NetworkLogRoute(val route: String) {
+    Main("network_main"),
+    Detail("network_detail"),
+}
