+++ feature/devmode/src/main/kotlin/com/wepli/feature/devmode/network/main/navigation/NetworkLogNavigation.kt
@@ -1,0 +1,22 @@
+package com.wepli.feature.devmode.network.main.navigation
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import com.wepli.feature.devmode.network.main.screen.NetworkLogDebugScreenRoute
+import com.wepli.feature.devmode.navigation.NetworkLogRoute
+
+fun NavController.navigateToNetworkLogMain() {
+    navigate(NetworkLogRoute.Main.route)
+}
+
+fun NavGraphBuilder.networkLogMainGraph(
+    navOnNetworkLogDetail: (Int) -> Unit,
+    navOnBack: () -> Unit
+) {
+    composable(
+        route = NetworkLogRoute.Main.route
+    ) {
+        NetworkLogDebugScreenRoute(navOnNetworkLogDetail = navOnNetworkLogDetail, navOnBack = navOnBack)
+    }
+}
